email sending and pressure sending

+++ Pressure_Monitoring_System_4_0_beta_copy_20231204091214/sensor.ino
@@ -2,7 +2,7 @@
 
 ADS1115 ADS(0x48);
 float f = 0;
-double pressure;
+float pressure;
 
 
 void setupADS()
@@ -32,8 +32,8 @@ int16_t readSensor()
 }
 
 
-double calcPressure(int16_t sens){
-  double sensVal=sens;
+float calcPressure(int16_t sens){
+  float sensVal=sens;
   pressure=(10.0/26400.0)*sens;
   // Serial.print("sensor: ");
   // Serial.print(val_0);
